revert move of lenses format helper functions to commonMain
